+++ NEWS.md
@@ -6,6 +6,7 @@
 - New root functions
 - Bulk density stored in soil object initialization
 - Advanced plant water pools
+- Belowground inputs restructured
 
 # Version 1.0.0
 - Reorganization of growth function